# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  getimages.HumpbackWhaleImage:
    properties:
      author_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      saw_at:
        type: string
      username:
        type: string
      whale_type:
        $ref: '#/definitions/getimages.WhaleType'
    type: object
  getimages.WhaleType:
    properties:
      conservation_status:
        type: string
      family:
        type: string
      genus:
        type: string
      id:
        type: string
      species_eng:
        type: string
      species_rus:
        type: string
    type: object
  getwhaletypes.WhaleType:
    properties:
      conservation_status:
        type: string
      family:
        type: string
      genus:
        type: string
      id:
        type: string
      species_eng:
        type: string
      species_rus:
        type: string
    type: object
  handler.getImagesResp:
    properties:
      next_page_url:
        type: string
      prev_page_url:
        type: string
      whale_images:
        items:
          $ref: '#/definitions/getimages.HumpbackWhaleImage'
        type: array
    type: object
  handler.getWhaleTypesResp:
    properties:
      whale_types:
        items:
          $ref: '#/definitions/getwhaletypes.WhaleType'
        type: array
    type: object
  handler.httpError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  handler.reqLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.reqRegister:
    properties:
      degree:
        type: string
      email:
        type: string
      password:
        type: string
      place_of_work:
        type: string
      rank:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  handler.respLogin:
    properties:
      is_scientist:
        type: boolean
      token:
        type: string
      username:
        type: string
    type: object
  handler.respRegister:
    properties:
      is_scientist:
        type: boolean
      token:
        type: string
      username:
        type: string
    type: object
...
host: localhost:80
info:
  contact:
    name: humpback-whale
    url: https://github.com/OkDenAl/humback-whale
  description: Humpback whale catalog - it is the service for recognize humpback whale
    and store it in catalog.
  title: Humpback whale catalog service
  version: "1.0"
paths:
  /private/whale/{whale_id}:
    delete:
      parameters:
      - default: Bearer <token>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Whale Image ID (UUID)
        in: path
        name: whale_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Whale ID format
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Whale image not found
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Удаляет наблюдение кита
      tags:
      - Наблюдение кита
  /private/whale/types:
    post:
      parameters:
      - default: Bearer <token>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Создаёт новый вид кита или обновляет существующий
      tags:
      - Вид кита
  /private/whale/types/{whale_type_id}:
    delete:
      description: Deletes a whale type record.
      parameters:
      - default: Bearer <token>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Whale Type ID (UUID)
        in: path
        name: whale_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Whale Type ID format
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Удаляет вид кита
      tags:
      - Вид кита
...
  /public/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: JSON
        in: body
        name: text
        required: true
        schema:
          $ref: '#/definitions/handler.reqLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.respLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Авторизация пользователя в системе
      tags:
      - Аутентификация
  /public/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: JSON
        in: body
        name: text
        required: true
        schema:
          $ref: '#/definitions/handler.reqRegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.respRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Регистрация пользователя в системе
      tags:
      - Аутентификация
...
  /public/whale/images:
    get:
      parameters:
      - description: Limit per page
        in: query
        name: limit
        required: true
        type: integer
      - description: Timestamp cursor (RFC3339Nano format) for pagination
        in: query
        name: cursor
        type: string
      - description: Filter by author username
        in: query
        name: username
        type: string
      - description: Filter by whale type id
        in: query
        name: whale_type_id
        type: string
      - description: Filter by start time (RFC3339Nano format)
        in: query
        name: start_time
        type: string
      - description: Filter by end time (RFC3339Nano format)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with whale images
          schema:
            $ref: '#/definitions/handler.getImagesResp'
        "400":
          description: Bad request (invalid parameters)
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not found (no images match filters)
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Получить список наблюдений с фильтрами
      tags:
      - Наблюдение кита
  /public/whale/types:
    get:
      description: Get a list of all available whale types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getWhaleTypesResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Получить список всех видов китов
      tags:
      - Вид кита
swagger: "2.0"